requirements/base.txt
django==1.11
django-annoying
pillow
pytils
django-extra-views
django-suit==0.2.26

